chore: cleanup header
