+++ src/components/Header/index.tsx
@@ -1,103 +1,14 @@
 import cn from 'classnames';
 import * as React from 'react';
-import { Head, withRouteData, withSiteData } from 'react-static';
 import Headroom from 'react-headroom';
+import { Head, withRouteData, withSiteData } from 'react-static';
 
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-
+import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { Link } from 'src/components/Link';
-import { Popup } from 'src/components/Popup';
-import { Portal } from 'src/components/Portal';
+import { Desktop } from './Desktop';
+import { Mobile } from './Mobile';
 
 export const headerHeightClass = 'h-20';
-
-const DropdownItem = (item, index) => {
-  return (
-    <Link
-      key={index}
-      to={item.href}
-      className={cn('flex px-3 text-black hover:opacity-75 cursor-pointer', {
-        'pt-6': index > 0,
-      })}
-    >
-      <div className={cn('flex items-center', item.subtitle ? 'w-12 mr-6' : 'w-8 mr-3')}>
-        {item.icon && (
-          <FontAwesomeIcon
-            className={cn(item.titleColor && `text-${item.titleColor}`)}
-            icon={item.icon}
-            size={item.subtitle ? '2x' : 'lg'}
-          />
-        )}
-      </div>
-
-      <div className="flex-1">
-        <div className={cn('text-lg font-bold', item.titleColor && `text-${item.titleColor}`)}>{item.title}</div>
-
-        {item.subtitle && <div className="text-base opacity-75">{item.subtitle}</div>}
-      </div>
-    </Link>
-  );
-};
-
-const HeaderDropdown = ({ width, title, links }) => {
-  if (!links || !links.length) {
-    return null;
-  }
-
-  return (
-    <Popup
-      width={width}
-      posX="center"
-      posY="bottom"
-      renderTrigger={attributes => (
-        <div className="flex select-none cursor-default text-white py-2 px-4 mx-2 font-semibold" {...attributes}>
-          <div className="flex-1 mr-2">{title}</div>
-        </div>
-      )}
-      renderContent={() => <div className="bg-white rounded-lg shadow-lg p-6">{links.map(DropdownItem)}</div>}
-    />
-  );
-};
-
-const HeaderButton = ({ title, href, icon, className }) => {
-  return (
-    <Link
-      key="2"
-      to={href}
-      className={cn("text-lg font-semibold py-2 px-4 ml-6 flex items-center border rounded text-white hover:text-white border-lighten-300 hover:border-lighten-500 bg-lighten-50 whitespace-no-wrap", className || '')}
-    >
-      {title} {icon && <FontAwesomeIcon icon={icon} className="ml-3" />}
-    </Link>
-  );
-};
-
-const Desktop = ({ items }) => {
-  return (
-    <div className="sm:hidden flex flex-1 justify-end items-center text-lg">
-      {items &&
-        items.length > 0 &&
-        items.map((item, index) => {
-          if (item.links && item.links.length) {
-            return <HeaderDropdown key={index} {...item} />;
-          }
-
-          if (item.isButton) {
-            return <HeaderButton key={index} {...item} />;
-          }
-
-          return (
-            <Link
-              key={index}
-              to={item.href}
-              className={cn("text-white hover:opacity-85 hover:text-white py-2 px-4 mx-2 font-semibold", item.className || '')}
-            >
-              {item.title}
-            </Link>
-          );
-        })}
-    </div>
-  );
-};
 
 export interface IHeaderLink {
   title: string;
@@ -109,104 +20,12 @@ export interface IHeaderLink {
 export interface IHeaderItem {
   title: string;
   href: string;
+  isButton?: boolean;
+  hideMobile?: boolean;
   className?: string;
   links?: IHeaderLink[];
-}
-
-export interface IMobileHeader {
-  items: IHeaderItem[];
-}
-
-class Mobile extends React.Component<IMobileHeader, { showMenu: boolean }> {
-  public state = {
-    showMenu: false,
-  };
-
-  public render() {
-    const { items } = this.props;
-    const { showMenu } = this.state;
-
-    const [main, ...extras] = items || [];
-
-    return (
-      <div className="hidden sm:flex flex-1 justify-end">
-        <FontAwesomeIcon
-          icon={['fas', 'bars']}
-          className="cursor-pointer ml-3 text-white"
-          size="2x"
-          onClick={() => this.setState({ showMenu: true })}
-        />
-
-        {showMenu && (
-          <Portal>
-            <div className="absolute pin z-10 flex flex-col">
-              <div className="relative m-4 pt-6 bg-white rounded" onClick={() => this.setState({ showMenu: false })}>
-                <div className="px-6">
-                  <div className="absolute pin-t pin-r p-4 flex items-center">
-                    <FontAwesomeIcon
-                      icon={['fas', 'times']}
-                      className="cursor-pointer text-grey"
-                      size="lg"
-                      onClick={() => this.setState({ showMenu: false })}
-                    />
-                  </div>
-
-                  {main && (
-                    <div className="text-md">
-                      <div className="pb-3 uppercase font-bold text-grey-darker">{main.title}</div>
-                      <div className="flex flex-wrap">
-                        {main.links &&
-                          main.links.map((product, index) => {
-                            return (
-                              <Link key={index} to={product.href} className="w-full flex items-center text-black py-4">
-                                {product.icon && (
-                                  <FontAwesomeIcon
-                                    className={cn(product.titleColor && `text-${product.titleColor}`)}
-                                    icon={product.icon}
-                                    // size="sm"
-                                  />
-                                )}
-
-                                <div className="flex-1 ml-3">
-                                  <div className={cn('font-bold', product.titleColor && `text-${product.titleColor}`)}>
-                                    {product.title}
-                                  </div>
-                                </div>
-                              </Link>
-                            );
-                          })}
-                      </div>
-                    </div>
-                  )}
-
-                  <div className="flex flex-wrap border-t mt-4 py-4 font-bold text-black text-md">
-                    {extras &&
-                      extras.length > 0 &&
-                      extras.map((item, index) => {
-                        if (item.links && item.links.length) {
-                          return item.links.map((link, linkIndex) => (
-                            <Link key={linkIndex} to={link.href} className="w-1/2 py-3">
-                              {link.title}
-                            </Link>
-                          ));
-                        }
-
-                        return (
-                          <Link key={index} to={item.href} className="w-1/2 py-3">
-                            {item.title}
-                          </Link>
-                        );
-                      })}
-                  </div>
-                </div>
-              </div>
-              <div className="flex-grow" onClick={() => this.setState({ showMenu: false })} />
-            </div>
-          </Portal>
-        )}
-      </div>
-    );
-  }
+  width?: number;
+  icon?: IconProp;
 }
 
 export interface IHeader {
@@ -233,19 +52,19 @@ class Header extends React.Component<IHeader, IHeaderState> {
     this.onUnfix = this.onUnfix.bind(this);
   }
 
-  onUnpin() {
-    this.setState({unpinned: true});
+  public onUnpin() {
+    this.setState({ unpinned: true });
   }
 
-  onUnfix() {
-    this.setState({unpinned: false});
+  public onUnfix() {
+    this.setState({ unpinned: false });
   }
 
   public render() {
     const { header, meta, color } = this.props;
     const { unpinned } = this.state;
     const bgClassName = unpinned ? `bg-${color || 'black'}` : '';
-    let headerItems = header.items || [];
+    let headerItems = (header && header.items) || [];
 
     if (headerItems.length) {
       if (unpinned) {
@@ -275,8 +94,14 @@ class Header extends React.Component<IHeader, IHeaderState> {
         <Head key="meta">
           <title>{meta && meta.title}</title>
         </Head>
-        <header key="header" className="absolute custom--z-100 pin-t pin-l pin-r">
-          <Headroom disableInlineStyles={true} downTolerance={65} onUnpin={this.onUnpin} onUnfix={this.onUnfix} className={cn(bgClassName)}>
+        <header key="header" className="absolute z-40 pin-t pin-l pin-r">
+          <Headroom
+            disableInlineStyles={true}
+            downTolerance={65}
+            onUnpin={this.onUnpin}
+            onUnfix={this.onUnfix}
+            className={cn(bgClassName)}
+          >
             <div className="container">
               <nav className={cn(headerHeightClass, 'flex items-center')}>
                 <Link to="/" className="text-white hover:opacity-75 hover:text-white text-2xl font-bold">
