+++ src/styles/app.scss
@@ -52,7 +52,7 @@ ul {
 
     &:nth-of-type(even) {
       background: #f5f7f9;
-      z-index: 50;
+      z-index: 30;
     }
 
     &:first-of-type {
@@ -1132,12 +1132,6 @@ pre.line-numbers > code {
 .line-numbers .line-highlight:before,
 .line-numbers .line-highlight:after {
   content: none;
-}
-
-
-// tailwind does not have z-100
-.custom--z-100 {
-  z-index: 100;
 }
 
 // Sticky header
